requirements.txt
Django==5.0.8
djangorestframework==3.15.2
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
stripe==9.12.0
requests==2.32.3
python-decouple==3.8
django-ratelimit==4.1.0
celery==5.3.4
redis==5.0.8
gunicorn==22.0.0
whitenoise==6.7.0